:sparkles: feat: add login modal view
Add modal view using auth0 api loginWithPopup function
