+++ src/components/Header/Header.styles.js
@@ -48,19 +48,6 @@ export const Drawer = styled(Stack)`
   }
 `;
 
-export const Overlay = styled(Box)`
-  background-color: rgba(84, 84, 84, 0.4);
-  display: ${(props) => (props.show ? 'block' : 'none')};
-  height: 200vh;
-  left: 0;
-  padding: 2rem;
-  position: fixed;
-  top: 0;
-  width: 150vw;
-  z-index: 1;
-  transition: 0.4s;
-`;
-
 export const LeftNav = styled(Stack)`
   flex-direction: row;
   width: 600px;
